// client/src/game/shared.ts
export const COMBAT_STATS = [
  'health',
  'stamina',
  'mana',
  'agility',
  'defense',
  'accuracy',
  'strength',
  'intelligence',
  'dexterity',
  'sneak',
  'awareness',
  'slash',
  'thrust',
  'spell',
  'oversize',
  'shoot',
] as const;
export type CombatStat = typeof COMBAT_STATS[number];
// ...
export interface IStat {
  level: number;
  xp: number;
  nextXp: number;
  points: number;
}
// ...
export const TASKS = ['routine', 'follow', 'dead'] as const;
export type TaskType = typeof TASKS[number];

export interface TrackAI {
  map: MapLocation;
  enemies: AIEnemy[];
}

export interface AIEnemy {
  id: string;
  movement: AIMovement;
  current: TaskType;
}

export interface AIMovement {
  velx: number;
  vely: number;
  dir: Direction;
  anim: PlayerAnimation;
  pos: { x: number; y: number };
}

export interface InitGame {
  map: MapLocation;
  users: IUser[];
  enemies: IEnemy[];
  loots: IUpdateLoot[];
}

export interface IPlayerMovement {
  name: string;
  dir: Direction | 'end';
  speed?: number;
  pos?: { x: number; y: number };
}

export const DIRECTIONS = ['up', 'down', 'left', 'right'] as const;
export type Direction = typeof DIRECTIONS[number];
// ...
export type PlayerAnimation =
  | PlayableAnimations
  | IdleDirection
  | MoveDirection;